math: implement most of glibc's __*_finite functions
These are supposed to be specialisations for speed, but these are just
faked. Some warnings were added too, if they return infinite values.

As a side effect of this change, scalbl is also now implemented.

As noted, not all functions are implemented; the big two blockers are an
implementation of j0l and y0l; I imagine Bessel functions aren't too
widely used, so I doubt that many things will want them. Someone (not
it) can implement them later.